[Internal] Bump Go SDK to v0.86.0 & Generate Latest TF Structs (#5092)
## Changes
<!-- Summary of your changes that are easy to understand -->
Bump Go SDK to v0.86.0 & Generate Latest TF Structs.

## Tests
<!--
How is this tested? Please see the checklist below and also describe any
other relevant tests
-->

- [ ] `make test` run locally
- [ ] relevant change in `docs/` folder
- [ ] covered with integration tests in `internal/acceptance`
- [ ] using Go SDK
- [ ] using TF Plugin Framework
- [ ] has entry in `NEXT_CHANGELOG.md` file

---------

+++ internal/providers/pluginfw/products/recipient_federation_policy/resource_recipient_federation_policy.go
@@ -11,6 +11,7 @@ import (
 	"github.com/databricks/databricks-sdk-go/apierr"
 	"github.com/databricks/databricks-sdk-go/service/sharing"
 	"github.com/databricks/terraform-provider-databricks/internal/providers/pluginfw/autogen"
+	pluginfwcommon "github.com/databricks/terraform-provider-databricks/internal/providers/pluginfw/common"
 	pluginfwcontext "github.com/databricks/terraform-provider-databricks/internal/providers/pluginfw/context"
 	"github.com/databricks/terraform-provider-databricks/internal/providers/pluginfw/converters"
 	"github.com/databricks/terraform-provider-databricks/internal/providers/pluginfw/tfschema"
@@ -40,7 +41,21 @@ type FederationPolicyResource struct {
 
 // FederationPolicy extends the main model with additional fields.
 type FederationPolicy struct {
-	sharing_tf.FederationPolicy
+	// Description of the policy. This is a user-provided description.
+	Comment types.String `tfsdk:"comment"`
+	// System-generated timestamp indicating when the policy was created.
+	CreateTime types.String `tfsdk:"create_time"`
+	// Unique, immutable system-generated identifier for the federation policy.
+	Id types.String `tfsdk:"id"`
+	// Name of the federation policy. A recipient can have multiple policies
+	// with different names. The name must contain only lowercase alphanumeric
+	// characters, numbers, and hyphens.
+	Name types.String `tfsdk:"name"`
+	// Specifies the policy to use for validating OIDC claims in the federated
+	// tokens.
+	OidcPolicy types.Object `tfsdk:"oidc_policy"`
+	// System-generated timestamp indicating when the policy was last updated.
+	UpdateTime types.String `tfsdk:"update_time"`
 }
 
 // GetComplexFieldTypes returns a map of the types of elements in complex fields in the extended
@@ -51,7 +66,9 @@ type FederationPolicy struct {
 // They must be either primitive values from the plugin framework type system
 // (types.String{}, types.Bool{}, types.Int64{}, types.Float64{}) or TF SDK values.
 func (m FederationPolicy) GetComplexFieldTypes(ctx context.Context) map[string]reflect.Type {
-	return m.FederationPolicy.GetComplexFieldTypes(ctx)
+	return map[string]reflect.Type{
+		"oidc_policy": reflect.TypeOf(sharing_tf.OidcFederationPolicy{}),
+	}
 }
 
 // ToObjectValue returns the object value for the resource, combining attributes from the
@@ -61,36 +78,98 @@ func (m FederationPolicy) GetComplexFieldTypes(ctx context.Context) map[string]r
 // interfere with how the plugin framework retrieves and sets values in state. Thus, FederationPolicy
 // only implements ToObjectValue() and Type().
 func (m FederationPolicy) ToObjectValue(ctx context.Context) basetypes.ObjectValue {
-	embeddedObj := m.FederationPolicy.ToObjectValue(ctx)
-	embeddedAttrs := embeddedObj.Attributes()
-
 	return types.ObjectValueMust(
 		m.Type(ctx).(basetypes.ObjectType).AttrTypes,
-		embeddedAttrs,
+		map[string]attr.Value{"comment": m.Comment,
+			"create_time": m.CreateTime,
+			"id":          m.Id,
+			"name":        m.Name,
+			"oidc_policy": m.OidcPolicy,
+			"update_time": m.UpdateTime,
+		},
 	)
 }
 
 // Type returns the object type with attributes from both the embedded TFSDK model
 // and contains additional fields.
 func (m FederationPolicy) Type(ctx context.Context) attr.Type {
-	embeddedType := m.FederationPolicy.Type(ctx).(basetypes.ObjectType)
-	attrTypes := embeddedType.AttributeTypes()
-
-	return types.ObjectType{AttrTypes: attrTypes}
+	return types.ObjectType{
+		AttrTypes: map[string]attr.Type{"comment": types.StringType,
+			"create_time": types.StringType,
+			"id":          types.StringType,
+			"name":        types.StringType,
+			"oidc_policy": sharing_tf.OidcFederationPolicy{}.Type(ctx),
+			"update_time": types.StringType,
+		},
+	}
 }
 
 // SyncFieldsDuringCreateOrUpdate copies values from the plan into the receiver,
 // including both embedded model fields and additional fields. This method is called
 // during create and update.
-func (m *FederationPolicy) SyncFieldsDuringCreateOrUpdate(ctx context.Context, plan FederationPolicy) {
-	m.FederationPolicy.SyncFieldsDuringCreateOrUpdate(ctx, plan.FederationPolicy)
+func (to *FederationPolicy) SyncFieldsDuringCreateOrUpdate(ctx context.Context, from FederationPolicy) {
+	if !from.OidcPolicy.IsNull() && !from.OidcPolicy.IsUnknown() {
+		if toOidcPolicy, ok := to.GetOidcPolicy(ctx); ok {
+			if fromOidcPolicy, ok := from.GetOidcPolicy(ctx); ok {
+				// Recursively sync the fields of OidcPolicy
+				toOidcPolicy.SyncFieldsDuringCreateOrUpdate(ctx, fromOidcPolicy)
+				to.SetOidcPolicy(ctx, toOidcPolicy)
+			}
+		}
+	}
 }
 
 // SyncFieldsDuringRead copies values from the existing state into the receiver,
 // including both embedded model fields and additional fields. This method is called
 // during read.
-func (m *FederationPolicy) SyncFieldsDuringRead(ctx context.Context, existingState FederationPolicy) {
-	m.FederationPolicy.SyncFieldsDuringRead(ctx, existingState.FederationPolicy)
+func (to *FederationPolicy) SyncFieldsDuringRead(ctx context.Context, from FederationPolicy) {
+	if !from.OidcPolicy.IsNull() && !from.OidcPolicy.IsUnknown() {
+		if toOidcPolicy, ok := to.GetOidcPolicy(ctx); ok {
+			if fromOidcPolicy, ok := from.GetOidcPolicy(ctx); ok {
+				toOidcPolicy.SyncFieldsDuringRead(ctx, fromOidcPolicy)
+				to.SetOidcPolicy(ctx, toOidcPolicy)
+			}
+		}
+	}
+}
+
+func (m FederationPolicy) ApplySchemaCustomizations(attrs map[string]tfschema.AttributeBuilder) map[string]tfschema.AttributeBuilder {
+	attrs["comment"] = attrs["comment"].SetOptional()
+	attrs["create_time"] = attrs["create_time"].SetComputed()
+	attrs["create_time"] = attrs["create_time"].(tfschema.StringAttributeBuilder).AddPlanModifier(stringplanmodifier.UseStateForUnknown()).(tfschema.AttributeBuilder)
+	attrs["id"] = attrs["id"].SetComputed()
+	attrs["id"] = attrs["id"].(tfschema.StringAttributeBuilder).AddPlanModifier(stringplanmodifier.UseStateForUnknown()).(tfschema.AttributeBuilder)
+	attrs["name"] = attrs["name"].SetOptional()
+	attrs["oidc_policy"] = attrs["oidc_policy"].SetOptional()
+	attrs["update_time"] = attrs["update_time"].SetComputed()
+
+	attrs["name"] = attrs["name"].(tfschema.StringAttributeBuilder).AddPlanModifier(stringplanmodifier.UseStateForUnknown()).(tfschema.AttributeBuilder)
+	return attrs
+}
+
+// GetOidcPolicy returns the value of the OidcPolicy field in FederationPolicy as
+// a sharing_tf.OidcFederationPolicy value.
+// If the field is unknown or null, the boolean return value is false.
+func (m *FederationPolicy) GetOidcPolicy(ctx context.Context) (sharing_tf.OidcFederationPolicy, bool) {
+	var e sharing_tf.OidcFederationPolicy
+	if m.OidcPolicy.IsNull() || m.OidcPolicy.IsUnknown() {
+		return e, false
+	}
+	var v sharing_tf.OidcFederationPolicy
+	d := m.OidcPolicy.As(ctx, &v, basetypes.ObjectAsOptions{
+		UnhandledNullAsEmpty:    true,
+		UnhandledUnknownAsEmpty: true,
+	})
+	if d.HasError() {
+		panic(pluginfwcommon.DiagToString(d))
+	}
+	return v, true
+}
+
+// SetOidcPolicy sets the value of the OidcPolicy field in FederationPolicy.
+func (m *FederationPolicy) SetOidcPolicy(ctx context.Context, v sharing_tf.OidcFederationPolicy) {
+	vs := v.ToObjectValue(ctx)
+	m.OidcPolicy = vs
 }
 
 func (r *FederationPolicyResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
@@ -98,10 +177,7 @@ func (r *FederationPolicyResource) Metadata(ctx context.Context, req resource.Me
 }
 
 func (r *FederationPolicyResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
-	attrs, blocks := tfschema.ResourceStructToSchemaMap(ctx, FederationPolicy{}, func(c tfschema.CustomizableSchema) tfschema.CustomizableSchema {
-		c.AddPlanModifier(stringplanmodifier.UseStateForUnknown(), "name")
-		return c
-	})
+	attrs, blocks := tfschema.ResourceStructToSchemaMap(ctx, FederationPolicy{}, nil)
 	resp.Schema = schema.Schema{
 		Description: "Terraform schema for Databricks recipient_federation_policy",
 		Attributes:  attrs,
@@ -114,12 +190,6 @@ func (r *FederationPolicyResource) Configure(ctx context.Context, req resource.C
 }
 
 func (r *FederationPolicyResource) update(ctx context.Context, plan FederationPolicy, diags *diag.Diagnostics, state *tfsdk.State) {
-	client, clientDiags := r.Client.GetWorkspaceClient()
-	diags.Append(clientDiags...)
-	if diags.HasError() {
-		return
-	}
-
 	var federation_policy sharing.FederationPolicy
 
 	diags.Append(converters.TfSdkToGoSdkStruct(ctx, plan, &federation_policy)...)
@@ -133,6 +203,12 @@ func (r *FederationPolicyResource) update(ctx context.Context, plan FederationPo
 		UpdateMask: "comment,oidc_policy",
 	}
 
+	client, clientDiags := r.Client.GetWorkspaceClient()
+
+	diags.Append(clientDiags...)
+	if diags.HasError() {
+		return
+	}
 	response, err := client.RecipientFederationPolicies.Update(ctx, updateRequest)
 	if err != nil {
 		diags.AddError("failed to update recipient_federation_policy", err.Error())
@@ -152,11 +228,6 @@ func (r *FederationPolicyResource) update(ctx context.Context, plan FederationPo
 func (r *FederationPolicyResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
 	ctx = pluginfwcontext.SetUserAgentInResourceContext(ctx, resourceName)
 
-	client, diags := r.Client.GetWorkspaceClient()
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
 	var plan FederationPolicy
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &plan)...)
 	if resp.Diagnostics.HasError() {
@@ -171,6 +242,13 @@ func (r *FederationPolicyResource) Create(ctx context.Context, req resource.Crea
 
 	createRequest := sharing.CreateFederationPolicyRequest{
 		Policy: federation_policy,
+	}
+
+	client, clientDiags := r.Client.GetWorkspaceClient()
+
+	resp.Diagnostics.Append(clientDiags...)
+	if resp.Diagnostics.HasError() {
+		return
 	}
 
 	response, err := client.RecipientFederationPolicies.Create(ctx, createRequest)
@@ -198,12 +276,6 @@ func (r *FederationPolicyResource) Create(ctx context.Context, req resource.Crea
 func (r *FederationPolicyResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
 	ctx = pluginfwcontext.SetUserAgentInResourceContext(ctx, resourceName)
 
-	client, diags := r.Client.GetWorkspaceClient()
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
 	var existingState FederationPolicy
 	resp.Diagnostics.Append(req.State.Get(ctx, &existingState)...)
 	if resp.Diagnostics.HasError() {
@@ -216,6 +288,12 @@ func (r *FederationPolicyResource) Read(ctx context.Context, req resource.ReadRe
 		return
 	}
 
+	client, clientDiags := r.Client.GetWorkspaceClient()
+
+	resp.Diagnostics.Append(clientDiags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
 	response, err := client.RecipientFederationPolicies.GetFederationPolicy(ctx, readRequest)
 	if err != nil {
 		if apierr.IsMissing(err) {
@@ -253,12 +331,6 @@ func (r *FederationPolicyResource) Update(ctx context.Context, req resource.Upda
 func (r *FederationPolicyResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
 	ctx = pluginfwcontext.SetUserAgentInResourceContext(ctx, resourceName)
 
-	client, diags := r.Client.GetWorkspaceClient()
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
 	var state FederationPolicy
 	resp.Diagnostics.Append(req.State.Get(ctx, &state)...)
 	if resp.Diagnostics.HasError() {
@@ -271,6 +343,12 @@ func (r *FederationPolicyResource) Delete(ctx context.Context, req resource.Dele
 		return
 	}
 
+	client, clientDiags := r.Client.GetWorkspaceClient()
+
+	resp.Diagnostics.Append(clientDiags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
 	err := client.RecipientFederationPolicies.Delete(ctx, deleteRequest)
 	if err != nil && !apierr.IsMissing(err) {
 		resp.Diagnostics.AddError("failed to delete recipient_federation_policy", err.Error())
